pyproject.toml
[tool.poetry]
name = "lievis"
version = "0.1.0"
description = ""
authors = ["Bastian Rieck <bastian@rieck.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9"
umap-learn = "^0.5.3"
matplotlib = "^3.6.2"

[tool.black]
line-length = 79


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
